Indicator led
As a testbed for the correct installation of the arduino id
I created an indicator light and uploaded to the arduino

Note: I had problem selecting the correct port. The reason
being that the usb board is only readable by root and dialout.

By adding my self to the dialout group with the command below

sudo usermod -G dialout -a <user>

this problem was fixed.